Last version

+++ accounts/src/main/java/com/hugo/accounts/model/Customer.java
@@ -17,10 +17,10 @@ public class Customer {
 
     @Id
     @GeneratedValue(strategy= GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
     private String email;
-    private Integer mobileNumber;
+    private Long mobileNumber;
     private LocalDate createdDate;
 
 }
